Path: Move more common methods to PathSurfaceSupport module
